Added UNIVERSAL_RULES Value
I believe we are on the same page on the non-invasive change, the previously proposed commit defaults to the current behavior, that is to say UNIVERSAL_RULES=true equals business as usual, UNIVERSAL_RULES=false equals scp compressed rules.

What I didn't account for is a securityonion.conf that does not have UNIVERSAL_RULES defined at all.  This is now fixed and sets UNIVERSAL_RULES to true by default and only if it exists in securityonion.conf and set to false will the added feature kick in.